backend/requirements.txt
# Core web framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP clients
httpx>=0.25.1
requests>=2.31.0

# AI/ML APIs
openai>=1.3.5
google-genai>=0.1.0
anthropic>=0.7.7

# Audio processing
pydub>=0.25.1
pedalboard>=0.8.7
librosa>=0.10.1
soundfile>=0.12.1

# Scientific computing
numpy>=1.24.3
scipy>=1.11.4

# Image/Video processing
opencv-python>=4.8.1
pillow>=10.1.0

# Web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Caching
redis>=5.0.1

# Async and utilities
aiofiles>=23.2.1
python-multipart>=0.0.6
websockets>=12.0
tenacity>=8.2.3
python-jose[cryptography]>=3.3.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0